requirements.txt
crewai
crewai_tools
crewai[tools]>=0.70.0
python-dotenv>=1.0.0
openai>=1.0.0
qdrant-client>=1.7.0
youtube-transcript-api>=0.6.0
pytube>=15.0.0
crewai[tools]>=0.70.0
python-dotenv>=1.0.0
openai>=1.0.0
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
faiss-cpu>=1.7.4
plotly>=5.18.0
pandas>=2.1.0
